Add comment and move test
