+++ pandas/tests/indexing/test_datetime.py
@@ -2,7 +2,6 @@ from datetime import datetime, timedelta
 
 from dateutil import tz
 import numpy as np
-import pytest
 
 import pandas as pd
 from pandas import DataFrame, Index, Series, Timestamp, date_range
@@ -314,20 +313,3 @@ class TestDatetimeIndex(object):
                                 columns=['value'],
                                 dtype=object)
         tm.assert_frame_equal(result, expected)
-
-    def test_getitem_with_datestring_with_UTC_offset(self):
-        # GH 24076
-        idx = pd.date_range(start='2018-12-02 14:50:00-07:00',
-                            end='2018-12-03 15:00:00-07:00', freq='1min')
-        df = pd.DataFrame(1, index=idx, columns=['A'])
-        result = df['2018-12-02 21:50:00+00:00':'2018-12-02 21:52:00+00:00']
-        expected = df.iloc[0:3, :]
-        tm.assert_frame_equal(result, expected)
-
-        # GH 16785
-        with pytest.raises(ValueError, match="Both date strings"):
-            df['2018-12-02 21:50:00+00:00':'2018-12-02 21:52:00+01:00']
-
-        with pytest.raises(ValueError, match="The index must be timezone"):
-            df = df.tz_localize(None)
-            df['2018-12-02 21:50:00+00:00':'2018-12-02 21:52:00+00:00']
